Fix root component metadata cache not invalidated in `beforeResolve`
This commit fixes #2480 by making sure that calls to methods which effectively mutate dependencies
hold by a configuration trigger mutation validation. This is done by wrapping the dependencies which
can be mutated by a proxy. This shouldn't be considered a final solution, as it has several issues.
First of all, using a proxy is a dirty workaround. This is done because there's tangling between
`core` and `dependencyMananagement`, and typically `ModuleDependency` lives into `core`, whereas
`MutationValidator` lives into `dependencyManagement`. Second, proxies have a performance cost (which
is going to prove to be a problem or not by the performance test suite).

A better fix would be to untangle everything and propagate the mutation validator.