Use a separate type for dependency attribute mutation
Gradle prevents changes to dependencies after a configuration has
been resolved or has been involved in resolution (as a parent configuration
or as a referenced project dependency).

Now that we are tracking mutation to Dependency attributes, changes that
were previously silently ignored should now fail the build. However, this
is a breaking change that affects the Spring DM plugin, and other plugins
that modify dependencies in a `beforeResolve` hook.

To avoid this breakage, we use a separate mutation type that is not (yet)
prohibited on resolved configurations. This will permit us to prohibit
this with a full deprecation cycle.