Add support for S3 multi-part uploads
If the file to be uploaded is larger than the threshold, the multi-
part path is used instead.
The request is first initiated, then each part is uploaded, and
finally the result is completed and the file is stiched back.
Only the client API interaction is tested at this point. Integration
tests to come.

Signed-off-by: Andreas Axelsson <axl@directivegames.com>