Enhanced FindBugs plugin to work with SpotBugs which is FindBugs successor.
FindBugs development has been discontinued. SpotBugs is the spiritual
successor of FindBugs, carrying on from the point where it left off
with support of its community.

The FindBugs plugin made an assumption that the FindBugs code was
present in an artifact named "findbugs-3.0.1.jar" in order to
extract the FindBugs version number to ensure it was compatible
with the current JDK.

In order to work with SpotBugs a new optional parameter
named "validateClasspath" which defaults to true.

Each updated file was also improved to use immutable parameters and
local variables whenever possible. This is common strategy in other
languages (Kotlin, Scala, JavaScript, etc.)

The documents were also updated to include some helpful tips with
using the FindBugs plugin (e.g. adding the "Find Security Bugs")

See "https://spotbugs.github.io/" and "http://find-sec-bugs.github.io/" for more information.