Disallow both --module-source-path and -sourcepath
Java 9 introduces a `--module-source-path` option which is mutually
exclusive with `-sourcepath` or `--source-path`. If they are both
provided, the sourcepath should be stripped and the
`--module-source-path` should remain.

If both have been specified explicitly by the user either as a
`optiions.compilerArgs` argument or, in the case of `sourcepath` as
the `options.sourcepath` property, then a warning should also be
emitted to let the user know that they did something wrong.