Don't apply fileOffset to streams for large S3 uploads
The fileOffset is only supposed to be used when you have a file,
for streams it will cause a double offset each time as the stream
has already been advanced by the previously uploaded segment.

This manifested itself in a failure to skip ahead in the stream
mid-through the upload as it reaches the end of the stream.

Signed-off-by: Andreas Axelsson <axl@directivegames.com>