Provide an API to get resolution errors when using lenient artifact views
This commit introduces a new API to collect resolution errors, when the resolution is lenient. Before,
only `Throwable`s were collected, resulting in very poor client side consumption. In particular, clients
have to resort on internal exception types and potentially parse the error messages to get context.

Instead, this commit adds a `ResolutionFailure` element, which carries enough context for the client
to deal with it. In particular, it will give access to a component identifier, or artifact identifier.
To make it simple to use, this commit also introduces a visitor pattern, that clients can implement in
order to collect only the information they need.

The idea is to add more visitors (into distinct interfaces, for backwards compatibilty) as we add more
context (for example, we could provide variant information in the context).