Make the verification of `Named` immutables stricter
This commit addresses review feedback: the check on `Named` was only checking that the user provided a type
extending `Named`, but they could very well provide a concrete type, which wouldn't be immutable. By making
sure we can only use an interface, then users can only use `Named` types which are actually not instantiable
directly. This won't prevent from passing a value which is a concrete type, but it will prevent declaring
attributes of mutable types.