Fix issue with `gradle idea` and Spring DM plugin
The Spring DM plugin modifies dependency excludes in a `beforeResolve`
hook. In certain circumstances this results in modifying a
dependency on an already-resolved configuration, which then fails
when the `idea` plugin resolves a second time.

This fixes the issue by ignoring changes to dependency excludes
post-resolve, as was the case in previous releases. Going forward
we should deprecate and eventually prohibit modification of
dependency attributes after a configuration has been resolved.