Make root field in test directory provider non-static
This field can have different values depending on wether test or
distribution directory provider is used. If both are used together,
they modify the same static field and the directories are mixed up.
Which was the case in all tooling API integration test. This has been
broken since the introduction of TestDistributionDirectoryProvider
(93df7b7).