Clean up retry rule util's debug output and add daemon context info
In case a flaky cross-version tests fail despite an attempt to retry:
With the latest changes to CI artifact publishing, the daemon logs
should all be available as separate artifacts. This change prints one
line with the daemon PIDs and Ports the test has (potentially) used
to help with failure investigation.