Make the options API public
Previously, the API for defining command line options for tasks was
internal. This commit makes the bare minimum of the API (namely the
Option and OptionValues annotations) public and deprecates the
existing, internal equivalents. The remainder of the existing
implementation remains part of the internal API and has been updated
to look for both the existing internal annotations and the new public
annotations.