Consolidate build cache performance tests
There is now only one class which contains all the tests. We now always
have the local cache enabled (push = true), since this is the default
use case. For the remote cache tests we clean the local cache after each
run.
We also added a new test which tests pushing to the remote build cache.

+review REVIEW-6556