Allow multi-version integration tests to set org.gradle.integtest.versions
This behaves similarly to TAPI tests that allow you to select a particular version
to test with.  This is an existing system property that was only used by
AbstractCompatibilityTestRunner.

Previously, you had to choose "latest" (which is the last element in the possible versions)
or "all" (which would test all versions) with org.gradle.integtest.multiversion.

For Play tests, I want to be able to test a single version.