Build operation queue wait thread can execute operations
This fixes an issue where when multiple tasks are running
concurrently and submitting work to the build operation queue,
performance can actually be worse worker leases and the
executor thread pool compete with each other.  This fixes the
issue by allowing the thread waiting for work to complete to
also execute some of that work.