Prevent parallel project configuration
When a Task resolves a Configuration that it didn't declare
as an input and another task does the same in parallel, then
both tasks can end up trying to configure a Project which was
not discovered as a dependency at configuration time. Configuring
projects in parallel is not safe.

This change guards access to the project form parallel resolvers
by acquiring the project lock first.