Increase default timeout for BlockingHttpServer
This class is used in long-running end-to-end tests which
often take a significant amount of time to even reach the
start of the work that the server is waiting for. This is
especially true on a heavily loaded build server.

The timeout is there as a last resort, to avoid builds hanging
forever. Making it a bit longer by default shouldn't hurt.