Touch cache entries for external resources when they are revalidated
Previously, the cache entry for an external resource retained it's original
timestamp, even after we had revalidated it against the source (ie. via HTTP HEAD).
This resulted in multiple HEAD revalidation requests being required, for a
cached resource that was requested multiple times in the same build.

This issue was made more apparent with the introduction of caching for
HTTP script plugins, but it had already been observed to affect the
caching of resources used by `ComponentMetadataSupplier`.

The fix supplied effectively 're-stores' the cache entry, forcing the
cache entry time to be updated.