Bump the number of ports allocated to a test worker
Currently we don't have any mechanism to allocate additional ports
if the provided range is exhausted, and some tests are hitting this
limit (20).
By bumping the limit to 50, this will still permit up to 320
test workers to be started on a machine, scaled down if there
are multiple CI agents running on the machine.