Add test for http build cache timeout
Fixing #868 via #2757 also introduces timeouts for connecting to an
HTTP build cache. We just add a test to verify this.
While writing the test I realized that we do not print the real cause
for failing with the HTTP backend when not printing the stack trace,
so I fixed that, too.