Remove programmatic control of Play reload trigger
Now, the user can decide how they want to interact with a Play app:
- To have the app eagerly built and redeployed on input file change
   `gradle --continuous runPlay`
- To have the app built and redeployed on HTTP request (with input file change)
   `gradle runPlay`

The Play plugin now always starts the deployment with `ChangeBehaviour.BLOCK`,
but a continuous build gate is only used when a reloadable deployment is discovered
at the end of the build.

Integration tests are yet to be updated for this behaviour.