Prevent file hashing from blocking on named pipes
Java will block when opening a named pipe (fifo), which causes problems
with tasks like `Sync`. If those tasks encounter any named pipes while
calculating snapshots of their source and destination filesets the
default file hasher attempts to open an input stream for each pipe in
order to compute a hash. This causes Gradle to hang indefinitely.

Named pipes are usually created on Unix platforms by the `mkfifo`
command, and a simple test in issue #2552 illustrates how to reproduce
the blocked I/O on any system where the `mkfifo` command is available.

Since the hash of a zero-length file is always the same and named pipes
always have length 0, it is possible to avoid opening a file input
stream when those files are encountered and instead return the hash of
the empty string.

- Also added new unit tests for the DefaultFileHasher class to expose
  the issue and confirm the existing behaviour is still correct.

Resolves: #2552