Rename `VersionSelector` to `VersionMatcher`
This refactor is introduced to reduce confusion: the selector corresponds to
the requested version, which can be an exact version number, a range, or
anything else. The matchers, on the other hand, match a selector. In other
words, the selector is a specification, while the matcher does the actual
work of matching.