Deprecate configure on demand
This feature was introduced to make configuration time in
large projects acceptable. This was back at a time when Gradle
was relatively inefficient. By now, the feature has lost a lot
of its appeal, due to many performance improvements.

Apart from the reduced usefulnes, it was also very hard to use
correctly and confused many users. E.g. the use of `allprojects {}`
 or `subprojects {}` negates the benefit of CoD, yet this is the
canonical way to cross-configure projects.

Going forward, we will ask users to provide profiling results if they
are experiencing slow configuration times. In most cases the problems
could be fixed in the user's build itself. In some cases we can make
some transparent improvement to Gradle.