Inline resolution actions
There are 2 cases where we iterate over conflicting modules, but the action is always the same. The first case is when a conflict is
discovered. We deselect the previous selection, but leave the resolution for later. Then, actual conflict resolution occurs and we
take the result of the conflict resolution and restart the resolution with it.

In practice, those actions are stateless: we don't need to create new actions for them, so this commit inlines them and makes them
part of the resolution state.