Rework `DefaultConflictResolutionResult`
This commit removes the need for storing `PotentialConflict` in `DefaultConflictHandler`. When the `resolveNextConflict` method
is called, we already know that it is an _actual_ conflict. The API made it unclear, by storing a `PotentialConflict` instance in
the conflict resolution result instance.

The only reason for this was to reuse the `withParticipatingModules` method. It's obviously simple enough that we can repeat it
and avoid the overhead of storing/removing potential conflicts in a map.