Changed the C++ plugins to make the link files for C++ libraries from a Maven repository visible to consumers at link time via dependency resolution.
The implementation is intentionally dumb and has plenty of edge cases. This implementation will get simpler and more robust as more metadata about each library is made available in the Maven repository and as the dependency resolution engine is able to consume this. In this change, the implementation works only when building on MacOs.