Fix corner case happening in nested closure scenario (#2888) (#2951)
https://github.com/gradle/gradle/issues/2888
As a flag to indicate the rootProject has been loaded,
DefaultGradle's rootProjectActions would be set to null after
projects are loaded, this is not clear enough. However, there's even
a bug: setting rootProjectActions to null should always happen,
no matter if rootProjectActions is empty. This PR fixes the
issue and adds a new boolean property projectsLoaded.