Stop imposing total ordering on UsageContexts
This prevents other software apart from Java libraries
to be published. This will be an issue both for native
libraries as well as when we open up the SoftwareComponent
API for extension. Some users already experimented with the
internal interfaces and we at least don't want to break them
for no good reason.