Remove design-docs folder
This has turned into a graveyard for ideas. It only serves
to confuse people at this point. We have found it more productive
to either

- use GitHub Epics and issues for smaller design questions
- use Google Docs for larger topics (e.g. native publishing)

These documents quickly go out of date once a feature is implemented.
They are not a replacement for good user and code documentation.

Many of the documents are about features that we never ended up
implementing. Having those documents still around might lock us
into a certain way of thinking about a problem. Instead we should
have a fresh look at it when we actually want to start working on it.