Improve version range selection
This commit fixes a couple of bugs in range selection. It refactors the code, making it clearer what the different states
of a module can be. In particular, the previous version included both `Conflict` and `Evicted`, but in practice there was
no difference in using one or the other!

Now the selection uses four different states (`Selected`, `Selectable`, `Evicted` and `Orphan`) which allows us to
discriminate between cases a node was evicted in conflict resolution, or when it was simplify preferred over another
possible choice.