Explode the beast into smaller pieces
This commit extracts all inner classes from `DependencyGraphBuilder`, as well as the builder itself, into its own package. The reason is that this class
had just became too big, and that there was a lot of shared state. By exploding it into well defined classes of their own, state can be locked and we can
effectively track where it is mutated. While not yet perfect, this should prevent headaches to future maintainers.