More efficient conflict registration
Instead of iterating over all conflicts (the list can grow big) and finding out if one of the conflicts
participants is the same as the current participants, we can build a reverse index (participant -> conflict)
and use it. This makes the check O(1) instead of O(n) (plus intersection computation).

Signed-off-by: Cedric Champeau <cedric@gradle.com>