Validate project name later
The name is initially set to the folder name although it might be
overridden later. If that is the case, we do not want to print a
deprecation warning.
If the folder name would be corrected, instead of accepting the
deprecated name, this would not be necessary. This breaking change
is planned for 5.0.