Revert change that broke Android testing
Android 2.3 uses the internal DurationFormatter class for its
test reporting. This has been fixed in Android 3.0, but we want
to keep Android 2.3 working a while longer. We will remove this
class in Android 5.0

This adds coverage for reporting on connected tests (the scenario that
broke). It does not add coverage for actually running the tests, since
that would require additional setup like an emulator.

This also adds coverage for Android 3.0 to ensure we don't break the
latest beta.