Store the task history in one cache (#3036)
Currently, Gradle is using two cache files for storing the task history: One for the file collection snapshots involved and another one storing the remainder of the task history.
Using a separate cache for file collection snapshots may cause inconsistent state as seen in #2827. Instead of having a different cache file for the snapshots we now have one cache for the task history including the snapshots.