Refactor populate to return working directory
In an attempt to implement locking of the git clone directory in
gradle/gradle-native#146, we realized that the directory locking
mechanism writes a lock file into the directory being locked. That
lock file prevents jGit from cloning into the directory because it
expects directories to be empty when they are the target of a clone
operation.

This change moves the responsibility for calculating the path to the
working directory into the VersionControlSystem and provides the
populate method with a parent directory which can be the target of
locking.

Part of gradle/gradle-native#146