Include simple test with locking
This is not sufficient to prove that pulling multiple
`VersionControlSpec`s in parallel to the same directory will be
safe. We still need to add some more sophisticated testing fixtures to
prove that.

This change also switches who has the responsibility for making the
"uniqueId()" for the `VersionControlSpec` globally unique. The
previous implementation wouldn't work given that the
`VersionControlSystem` would always be a
`ThreadSafeVersionControlSystem` no matter what the delegate
`VersionControlSystem` would be.

Part of gradle/gradle-native#146