Improve binary compatibility check report
The report now corresponds to what we had in the JDiff reports before:
- It also shows changes to Incubating API
- It excludes some more redundant change types for better readability

The report generator now also reports (and fails) for missing @since
tags. This is implemented by checking the source of a class that
includes new API elements suing the 'com.github.javaparser'.

Only changes to non-incubating public API need to be accepted
explicitly in accept-public-api-changes.json.