Separate concerns of merging and registering auto-applied plugins
The AutoAppliedPluginHandler is responsible for merging the user's
requests with the ones supplied by the AutoAppliedPluginRegistry.
It can do so in a generic way, independent of the build scan use case.

The registry on the other hand is currently hard-coded to only handle
the scan plugin. It could later be replaced with a pluggable implementation.

We still need to refactor `PluginRequest`, so it doesn't take a file name
and line number directly, but instead some more generic thing. Auto-applied
plugins don't have a line number or source file.