Use a `List` for outgoing edges and unattached dependencies
This commit changes the type of outgoing edges from `Set` to `List`. This can be done because the only place
where we actually add edges already makes sure that they are unique. As such, we don't need to compute a hashcode
or compare edges when added to the list. Profiling showed that ~4% of graph building was spent just here, for
the "exclude rule merging" scenario.

The same reasonaing applies for unattached dependencies: the rationale is that a single edge should only be
added once, so we don't need to check for unicity.