Build Gradle with latest nightly
This was required because our current API compatibility check uses the version
of the currently executing Gradle as a baseline. So when building with a pre-release
version of 4.3, then check requires every new API method to be annotated with `@since 4.3`.

This should probably be fixed properly, by using the version number from `version.txt`
for this check.