Parse error output to check for build failure in ForkingGradleHandle
ForkingGradleHandle was using the process exit value to determine if
a integTest build had failed. However, we've had numerous examples of
an exitValue == 0 where a failure is expected, and the error output
indicates failure.

To workaround this flakiness, we now use also check error output as
an indicator of build failure.

Fixes gradle/gradle-private#787