Remove support for `optional` in the DSL
The rationale behind this change is that we're not going to offer "optional" as a first-class concept. What we support is understanding
optional dependencies as found in POM files. However, optional dependencies will be better modelled using some kind of "features" declaration.
In other words, a dependency is often optional because it only needs to be brought in if you are using a particular feature of
a library. For example, "if you are going to use mysql as the backend, include the mysql dependency". But that's not really an optional
dependency: it's something that needs to be there when compiling the provider, but only for some consumers.