Fix xtend plugin smoke test
The introduction of optional dependencies broke the XTend plugin smoke test, because of a very strange
problem: now, the `org.eclipse.core.runtime` dependency has its optional dependency on `javax.inject`
used during dependency resolution, because it's also brought transitively, but as a hard dependency,
by Guice. However, the build failed with a spurious error, saying that no `javax.inject` version
matched the range `[1.0.0,)`. Indeed, there's no such `1.0.0` version, there's only version `1`. The
deal was to understand where the range `[1.0.0,)` came from, because the POM file for the Eclipse
dependency on Maven Central had `[1,)`, which would match as expected. The outcome was that the
dependency metadata on JCenter was different! On JCenter it seems that the range has been rewritten
to `[1.0.0,)`. So this commit fixes the problem by using `mavenCentral()` instead of `jcenter()`.