Ignore deprecation warnings in toolingApiBuilders project
Since we currently still compile against Java 7, we are affected
by a JDK bug that emits deprecation warnings when implementing a
deprecated interface, even when a SuppressWarnings annotation is used.
This was fixed in Java 8, so we should revert this change once we switch
to Java 8 as our minimum version.

See https://bugs.openjdk.java.net/browse/JDK-6480588