Only search for swiftc and Clang on macOS
platformTest tries to test all "supported" tool chains on a given OS
We assumed that if the OS was not Windows, we must be on Linux and
we must support GCC/Clang/swiftc.

On macOS, GCC is usually just a wrapper around Clang, so we detect
GCC as missing.  This causes tests to fail in platformTest because
we cannot find a GCC tool chain.