Remove `outputType` as input to native compilation
The manually added `outputType` input to native compilation tracks
- the target platform (os/architecture)
- the type of the toolchain (Clang/Gcc etc.)
- the name of the toolchain (arbitrary selected name)
- the OS of the toolchain

We already track the target platform and the type/version of the compiler
in use separately in CompilerVersion.
The OS of the toolchain should not influence the result of the
compilation. If it would, this could cause then problems for caching.

We decided not to track the OS of the toolchain for now, since
- if there is a problem when using the build cache we want to know about
  it
- there is an easy workaround if the problem arises: add the OS as an
  input to the compile task in the build script

The name of the toolchain should not influence the output at all.