Allow for gracefully completely disabling build scans when Gradle knows its not compatible with the scan plugin version.
This is driven by scan functionality currently not working in the presence of source mappings. Previously, the scan plugin was reaching through internal and public incubating API, making it public API in effect, in order to work out if there were VCS mappings. Instead, we now use our existing insulation layer to convey this.

Moreover, for future situations like this, it is now possible for Gradle to tell the scan plugin that it should not try and do its thing. Had this already existed, we could have used it here. Said another way, we can use this to avoid hard failures when we know something about a new version of Gradle makes it incompatible with existing scan plugin versions, which we are likely to know before we release the version of Gradle in question.