Reproduce store-first bug
Normally, the first request to the local cache is a `DirectoryBuildCacheService.load()`, which initializes the cache, so when we hit it with the first store, the directory is already there and prepared.

However, if the task is re-run (via `--rerun-tasks`, or because it's marked as `upToDateWhen()`), then Gradle won't load any results from cache. In this case the first access to the local cache is a store:

- we first create the ".part" file inside the build-cache directory in `DirectoryBuildCacheService.allocateTempFile()`,
- the cache directory is then initialized (and all content deleted, including the ".part" file) when we hit `DirectoryBuildCacheService.store()`
- we try to move the (now deleted) ".part" file into the cache, which promptly fails