Support `RegularFileProperty` as artifact in publishing plugins
A `RegularFileProperty` that is constructed by Task.newOutputFile() retains
a reference to the producing task. This commit allows these properties to
be used directly to construct a publication artifact for `maven-publish` and
`ivy-publish` plugins, without the need to directly wire in the task.

It is now easier to construct a publication without requiring the use of
`AbstractArchiveTask` for all artifacts produced.