Do not consider module file to determine main artifact or packaging
Currently, the .module metadata file is treated like a regular artifact in
maven publishing. In the longer term we should provide special treatment for
this artifact, much like we do for POM files.

In the meantime, we simply ignore this artifact when attempting to
determine the packaging or main artifact for a Maven publication.