Allow for gracefully completely disabling build scans (#3291)
This is driven by scan functionality currently not working in the presence of source mappings. Previously, the scan plugin was reaching through internal and public incubating API, making it public API in effect, in order to work out if there were VCS mappings. Instead, we now use our existing insulation layer to convey this.