Use a `boolean` instead of `ThreadLocal`
This changes the code in `ConfigureDelegate` to use a `boolean` instead
of a `ThreadLocal<Boolean>`. The rationale for this change is that the current
code leaks memory. On some scenarios, after a single build run, the thread
local map contains more than 60MB of memory just to retain booleans that will
never ever be used.

Instead, we're now using an `boolean`. The semantics are going to be
different, but in any case if code is called concurrently, all bets are off:
this code is supposed to be called only during configuration, and configuration
is single-threaded, so it _should not_ be called concurrently. However, we've
now made it explicit that this class is not thread safe.