Use an `AtomicBoolean` instead of `ThreadLocal`
This changes the code in `ConfigureDelegate` to use an `AtomicBoolean` instead
of a `ThreadLocal<Boolean>`. The rationale for this change is that the current
code leaks memory. On some scenarios, after a single build run, the thread
local map contains more than 60MB of memory just to retain booleans that will
never ever be used.

Instead, we're now using an `AtomicBoolean`. The semantics are going to be
different, but in any case if code is called concurrently, all bets are off.