Replace 'client modules' and '@jar' usages with metadata rules
Note that this is a first step that reproduces exactly the same
dependencies as the build had before the change. The intention why
dependencies where excluded or changed before is lost for most cases.
Some changes are most likely not even intended.
Each rule should be checked and and it should be clarified why it
is required (or it should be removed).