Split `VersionConstraint` into its mutable and immutable counterparts
The mutable part is used when configuring the build, while the immutable one is used at resolution time.
This pushes the version scheme selector to the descriptor factory, and removes the need for using it
in the repository chain resolver. This is only a first step, the string version of the module version
selector has yet to be removed, but it requires more changes that are in a separate story.