Make task dependencies cheaper
Every task depends on its own inputs. Until
now this was modeled by adding these inputs
to the set of task dependencies. This forces
creation of a set even when this will be the
only entry. It also means that the user can
overwrite it and create a task which no longer
depends on its own inputs.

This change introduces a specialized dependency
object that is aware of the task's inputs. This
avoids unnecessary garbage and makes sure that
the user cannot overwrite it.