Set NAME_ONLY path sensitivity for header dependencies
If we build this project:

```
// main.cpp

#include "a.h"
#include "b.h"

// a.h
class A {}

// b.h
class B : A {}
```

And then do:

```
mv a.h tmp
mv b.h a.h
mv tmp b.h

```

And rebuild, the compile tasks will be up-to-date when it should be a
 compilation failure. This is because we marked the headers with NONE
PathSensitivity. Now, we changed it to NAME_ONLY, that should be enough.