Avoid Maven repository lookups for unsupported notations
A Maven repository always requires a group, name and version to
be specified. Asking without them will result in either a 400 or 404,
depending on the artifact server. This creates unnecessary network traffic,
so instead we don't even ask and report the module as missing.

This can happen when a user mixes a flatDir repository and a Maven
repository. Modules in the flatDir repository don't have a group.
Depending on the order of repositories, Gradle will end up asking the
Maven repository for things that can never be found there.