Improve test fixtures for resolution
This commit introduces new test fixtures (built on top of the old ones) for dependency resolution, allowing
to setup the repository layout easily, as well as declaring which modules are going to be requested. It
will appropriately setup the expectations on metadata and artifacts based on the contextual execution, that
is to say if `--parallel` is on, and if Gradle module metadata is expected.