Add support for resolving strict dependencies from an Ivy repository
So far strict dependencies were only supported when resolving from a Maven repository. This commit adds
the necessary infrastructure to make it work on Ivy repositories too. It's worth noting that similarly
to Maven, as soon as Gradle metadata is used, variants from the original Ivy metadata file are "ignored",
and it's a lossy conversion.

This commit is a "make it work". There's still missing test coverage, and there's redundant code due
to the replication of what has been setup for the Maven repositories. Subsequent commits will fix that.