Let empty version selectors match anything
So far the empty version selector only matched an
empty version. However, what it really means is that
the user didn't care what this selector chooses and
wants the version to be recommended by something else.
Until now this would lead to conflict resolution, where
the empty version would be upgraded to a non-empty one.

Instead, the empty version is now replaced by a non-empty
one during selection and doesn't lead to conflict resolution
anymore. This leads to a more natural resolution result for
the BOM import case.