Reworked incremental native compile dependency analysis so that less state is persisted, and to separate the state that is collected for a source file and the state that is collected for a header file.
This change also starts to tease apart the work of calculating the set of header files for a source file from the work of deciding whether a source file needs to be recompiled. This will allow each of these things to be done in separate tasks, so that header file calculation can happen in the depend task and the up-to-date check can happen in the compile task without both tasks redoing the same work.