Revert "Prevent enabling experimental features after first use"
This reverts commit ecfc00145efae61fa6abfabbc67d3cdc2183be17.

Making this flag immutable after it has been read was problematic for the `CppPlugin`
which depends on the new module metadata format and sets the flag when it is
applied. If a build needs to read the flag for any reason prior to plugin application
(e.g. resolving plugins from a repo) then appling the `CppPlugin` will fail.