Reuse cached snapshots of unfiltered trees
It is quite common for one task to generate some output
and another task to take a filtered version of that as an
input. Up until now we would snapshot the directory again,
even though the generating task put a snapshot of its full
output into the in-memory cache.

Instead, we now check whether we already have a snapshot of
the unfiltered contents of a directory and filter it in memory
instead of walking the file system again. This significantly
reduces up-to-date build times for Android projects, since
the Android plugin uses plenty of filtered trees.