Fix local cache concurrency (#3540)
The problem was that the local directory cache did not prevent reads overlapping with writes in-process. The assumption was that writes would be atomic, but in practice they are unfortunately not.

With this commit we now use a read-write lock to make sure reads only overlap other reads, and writes are executed in isolation. Both reads and writes happen behind an inter-process lock, so there's always just one process reading from or writing to the build cache directory.