Get system headers for language
Gcc returns different system headers depending on the detected language
of the source file to compiler. For detecting system headers we were
using stdin as an input file - so the compiler always used C system
headers.
This is not correct. When compiling C++ files we should detect the C++
system headers, too.