Contextualize ConfigurationMetadata generated for Ivy/POM
DefaultDependencyDescriptor represents an _ivy_ dependency, which can
apply to many different configurations (due to the way Ivy descriptor files
are structured). With this change, we wrap any ConfigurationMetadata created
for these types, so that they only provide attributes that are relevant to
the defined configuration.

This change allows us to remove the `fromConfiguration` parameter from 2 methods
on DependencyMetadata, since the `fromConfiguration` is known by the dependency
instance itself.

The `fromComponent` and `configurations` parameters have not yet been removed.