Implement smarter strategy to resolve sources
This commit changes the way we find sources, by only looking into the appropriate repository. If the
version is a released version, then we will only look into the `distributions` repository. If it's
a snapshot, then we will look into the `distributions-snapshot` repository.

However, it's possible that this project depends on a Kotlin DSL custom distribution for which the
sources haven't been published. So to make it possible to resolve, we use a version range, ending
with the current version, and starting with the previous minor.