Remove trailing whitespace in CONTRIBUING.md
Run `git show -R <commit>` to see the removed trailing whitespace. Some
text editors and IDEs are configured to automatically remove that
whitespace, so it's better to remove it now than to leave it and risk
having the whitespace be removed in a commit that is not about removing
trailing whitespace.