Extracted `ExcludeMetadata` out of `Exclude`
`Exclude` represents an Ivy representation of an exclude rule, which
may apply to different configurations. In Gradle, an Exclude is attached to a single
configuration or a single dependency, so `Exclude.getConfigurations()` didn't make
sense.

With this change, `ExcludeMetadata` represents the portion of an exclude that actually
matters when resolving, pushing `Exclude` down to Ivy/Maven specific classes.