Centralize logic for mapping Ivy/Maven model to ConfigurationMetadata
Previously, the logic for mapping dependency declarations to configurations
was encoded in DefaultConfigurationMetadata, and was shared by Maven and Ivy subtypes.

With this change, this logic is in the `[Ivy|Maven]ModuleResolveMetadata` types
with simpler logic being used for the Maven case.
For consistency, the filtering of exclude rules was also moved out of
IvyConfigurationMetadata into the creator.