Introduce `alwaysProvidesMetadataForModules` on repositories
This commit adds the notion of repository content filter, which is there to optimize lookup
of modules. Precisely, it adds the `alwaysProvidesMetadataForModules` flag, which, when set
to `true`, disables the legacy Maven/Ivy lookup of artifacts (jar) whenever a module metadata
file is missing. When this flag is set to true, it means that it's not worth checking for a
`jar` when the corresponding `pom.xml` or `ivy.xml` file is not found. This avoids unnecessary
network requests when a module is never going to be found.

Rather than just introducing a flag on repositories, this commit also introduces a new interface,
named `RepositoryContentFilter`, which allows configuring what is expected to be found in a
repository. As of this commit, the only possible configuration flag is this always provides
metadata flag. But it is expected to find more configuration in the future, to, typically,
prevent any kind of remote lookup whenever we can be sure that a module is **not** going to
be found in a specific repository.