Prepare infrastructure for pluggable module metadata sources
This commit replaces the `contentFilter { ... }` block with a `metadataSources { ... }` block (currently not used)
which will let the user declare what metatadata sources are available in a specific repository. In practice, a
metadata source is a way to get metadata in a repository, for example, a POM file, an Ivy file, or Gradle metadata.